refactor: update internal file structure
