+++ src/App.tsx
@@ -1,15 +1,16 @@
 import { BrowserRouter, Routes, Route, Link } from 'react-router-dom';
-import Home from './pages/homePage';
-import './styles/app.css';
+import Home from './pages/HomePage';
 
 export default function App() {
   return (
     <BrowserRouter>
-      <main className="container">
+      <div className="container">
         <Routes>
-          <Route path="/" element={<Home />}></Route>
+          <Route path="/" element={<Home />}>
+            {/* <Route path="/" element={(props) => <Details {...props} />} /> */}
+          </Route>
         </Routes>
-      </main>
+      </div>
     </BrowserRouter>
   );
 }
