+++ src/main.tsx
@@ -1,17 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import ErrorBoundary from './react-components/errorBoundary';
+import ErrorBoundary from './react-components/ErrorBoundary/errorBoundary';
+import ErrorMessage from './react-components/ErrorMessage/ErrorMessage';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <ErrorBoundary
-      fallback={
-        <div className="error-message">
-          Error has been occured. Try again later or reload the page
-        </div>
-      }
-    >
+    <ErrorBoundary fallback={<ErrorMessage />}>
       <App />
     </ErrorBoundary>
   </React.StrictMode>
