+++ src/react-components/Details/Details.css
