+++ src/react-components/Details/types.ts
@@ -1,0 +1,6 @@
+import { IPlanet } from '../../utils/GeneralTypes';
+
+export interface IDetailsProp {
+  info: IPlanet;
+  onClose: () => void;
+}
