+++ src/react-components/ErrorBoundary/ErrorBoundary.tsx
@@ -1,13 +1,5 @@
 import React, { Component, ReactNode } from 'react';
-
-interface ErrorBoundaryProps {
-  fallback: ReactNode;
-  children: ReactNode;
-}
-
-interface ErrorBoundaryState {
-  hasError: boolean;
-}
+import { ErrorBoundaryProps, ErrorBoundaryState } from './types';
 
 export default class ErrorBoundary extends Component<
   ErrorBoundaryProps,
