+++ src/react-components/ErrorBoundary/types.ts
@@ -1,0 +1,10 @@
+import { ReactNode } from 'react';
+
+export interface ErrorBoundaryProps {
+  fallback: ReactNode;
+  children: ReactNode;
+}
+
+export interface ErrorBoundaryState {
+  hasError: boolean;
+}
