+++ src/react-components/ErrorButton/ErrorButton.css
@@ -1,0 +1,15 @@
+.error-button{
+    margin-left: 30px;
+    cursor: pointer;
+    height: 30px;
+    width: 100px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border: 1px solid black;
+}
+
+.error-button:hover{
+    background-color: rgba(1,1,1, 0.5);
+    color: white;
+}
