+++ src/react-components/ErrorMessage/ErrorMessage.css
@@ -1,0 +1,9 @@
+.error-message{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-size: 24px;
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    color: red;
+    height: 100vh;
+}
