+++ src/react-components/ErrorMessage/ErrorMessage.tsx
@@ -1,0 +1,9 @@
+import './ErrorMEssage.css';
+
+export default function ErrorMessage() {
+  return (
+    <div className="error-message">
+      Error has been occured. Try again later or reload the page
+    </div>
+  );
+}
