+++ src/react-components/Header/Header.css
@@ -29,13 +29,3 @@
     background-color: rgba(1,1,1, 0.5);
     color: white;
 }
-
-.error-message{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 24px;
-    font-family: Verdana, Geneva, Tahoma, sans-serif;
-    color: red;
-    height: 100vh;
-}
