+++ src/react-components/Pagination/types.ts
@@ -1,0 +1,5 @@
+export type PaginationProps = {
+  currentPage: number;
+  totalPages: number;
+  onPageChange: (page: number) => void;
+};
