+++ src/react-components/SearchResults/SearchResults.css
@@ -4,6 +4,16 @@
     box-sizing: border-box;
 }
 
+.loader{
+    font-size: 34px;
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: rgb(174, 174, 237);
+    height: 85vh;
+}
+
 .no-results{
     font-size: 34px;
     font-family: Verdana, Geneva, Tahoma, sans-serif;
