+++ src/react-components/SearchResults/types.ts
@@ -1,0 +1,6 @@
+import { IPlanet } from '../../utils/GeneralTypes';
+
+export type SearchResultProps = {
+  results: IPlanet[];
+  currentPage: number;
+};
