+++ src/react-components/SelectElement/types.ts
@@ -1,0 +1,3 @@
+export interface ISelectElements {
+  onItemsChange: (num: number) => void;
+}
