--- src/styles/app.css
@@ -1,9 +1,0 @@
-.loader{
-    font-size: 34px;
-    font-family: Verdana, Geneva, Tahoma, sans-serif;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: rgb(174, 174, 237);
-    height: 85vh;
-}
